address more reviewer comments
