+++ bundles/core/org.eclipse.smarthome.core.test/src/test/java/org/eclipse/smarthome/core/library/types/QuantityTypeTest.java
@@ -31,7 +31,6 @@ import org.junit.Before;
 import org.junit.Test;
 
 import tec.uom.se.quantity.QuantityDimension;
-import tec.uom.se.unit.Units;
 
 /**
  * @author Gaël L'hopital - initial contribution
@@ -231,18 +230,6 @@ public class QuantityTypeTest {
     @Test(expected = ArithmeticException.class)
     public void testDivide_Zero() {
         new QuantityType<>("4 m").divide(QuantityType.ZERO);
-    }
-
-    @Test
-    public void testPpm() {
-        QuantityType<Dimensionless> ppm = new QuantityType<>("500 ppm");
-        assertEquals("0.05 %", ppm.toUnit(Units.PERCENT).toString());
-    }
-
-    @Test
-    public void testDb() {
-        QuantityType<Dimensionless> ratio = new QuantityType<>("100");
-        assertEquals("20.0 dB", ratio.toUnit("dB").toString());
     }
 
     @Test
